Use latest beta pkg, remove all patterns except data-list
