+++ .circleci/config.yml
@@ -1,0 +1,32 @@
+version: 2
+jobs:
+  build:
+    working_directory: ~/angular-design-patterns
+    docker:
+      - image: circleci/node:12.6
+    steps:
+      - checkout
+      - restore_cache:
+          keys:
+            - v2-dependencies-{{ checksum "yarn.lock" }}
+      - run:
+          name: Install Dependencies
+          command: |
+            yarn install --frozen-lockfile
+      - save_cache:
+          name: Save Cache
+          paths:
+            - node_modules
+          key: v2-dependencies-{{ checksum "yarn.lock" }}
+      - run:
+          name: Prettier Check
+          command: |
+            yarn prettier:check
+      # - run:
+      #       name: Lint
+      #       command: |
+      #           yarn lint
+      # - run:
+      #       name: Unit Tests
+      #       command: |
+      #           yarn test
