+++ package.json
@@ -24,7 +24,7 @@
         "@angular/platform-browser": "~9.1.11",
         "@angular/platform-browser-dynamic": "~9.1.11",
         "@angular/router": "~9.1.11",
-        "@pxblue/angular-components": "2.0.0-beta.8",
+        "@pxblue/angular-components": "2.0.0-beta.9",
         "@pxblue/angular-themes": "^5.0.0-beta.8",
         "@pxblue/colors": "^3.0.0",
         "@pxblue/icons": "^1.0.25",
