+++ src/app/app.module.ts
@@ -16,34 +16,23 @@ import { MatListModule } from '@angular/material/list';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatTabsModule } from '@angular/material/tabs';
-import { DataListComponent } from './pages/data-list/data-list.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { FormValidationComponent } from './pages/form-validation/form-validation.component';
-import { ActionListComponent } from './pages/action-list/action-list.component';
-import { AppBarComponent } from './pages/app-bar/app-bar.component';
 import { DrawerLayoutModule, DrawerModule, EmptyStateModule } from '@pxblue/angular-components';
 import { PlaceholderComponent } from './pages/placeholder/placeholder.component';
-import {UnCamelCasePipe} from "./shared/un-camel-case.pipe";
-import {ObjectToList} from "./shared/key-value.pipe";
+import {DataListModule} from "./pages/list/data-list/data-list.module";
 
 @NgModule({
     declarations: [
         AppComponent,
-        DataListComponent,
-        FormValidationComponent,
-        ActionListComponent,
-        AppBarComponent,
         PlaceholderComponent,
-      UnCamelCasePipe,
-      ObjectToList
     ],
     imports: [
+      DataListModule,
         DrawerLayoutModule,
         DrawerModule,
         BrowserAnimationsModule,
         BrowserModule,
         AppRoutingModule,
-        FlexLayoutModule,
         MatToolbarModule,
         MatIconModule,
         MatCardModule,
