--- src/app/pages/action-list/action-list.component.html
@@ -1,35 +1,0 @@
-<mat-toolbar data-cy="pxb-toolbar" color="primary">
-    <h1>Action list</h1>
-    <div fxFlex style="margin: auto;"></div>
-    <button data-cy="delete-all" mat-icon-button (click)="onRemoveAll()"><mat-icon>delete</mat-icon></button>
-    <button data-cy="add item" mat-icon-button (click)="onAddItem()"><mat-icon>add</mat-icon></button>
-</mat-toolbar>
-<mat-nav-list *ngIf="data.length > 0">
-    <mat-list-item *ngFor="let i of data" [ngClass]="{ selected: isSelected(i) }" (click)="onSelected(i)">
-        <div mat-line>
-            {{ i.name }}
-        </div>
-        <p mat-line>
-            {{ i.details }}
-        </p>
-        <button
-            data-cy="action-menu"
-            mat-icon-button
-            #menuTrigger="matMenuTrigger"
-            [matMenuTriggerFor]="menu"
-            (click)="menuTrigger.openMenu()"
-        >
-            <mat-icon>more_vert</mat-icon>
-        </button>
-        <mat-menu #menu="matMenu">
-            <button data-cy="remove" mat-menu-item (click)="onRemoveItem(i)">Remove</button>
-            <button data-cy="details" mat-menu-item>View Details</button>
-        </mat-menu>
-    </mat-list-item>
-</mat-nav-list>
-<div *ngIf="data.length === 0" style="margin: 15px;">
-    <h1>No Items found</h1>
-    <button data-cy="add-item" mat-raised-button (click)="onAddItem()">
-        Add An Item
-    </button>
-</div>
