--- src/app/pages/action-list/action-list.component.scss
@@ -1,7 +1,0 @@
-.selected {
-    background-color: #cccccc;
-}
-
-mat-nav-list {
-    padding-top: 0px;
-}
