--- src/app/pages/app-bar/app-bar.component.html
@@ -1,59 +1,0 @@
-<div
-    #wrapper
-    class="eaton mat-typography"
-    style="overflow-x: hidden;"
-    fxFlexFill
-    cdkScrollable
-    (scroll)="onScroll($event)"
->
-    <mat-toolbar
-        color="primary"
-        [ngClass]="{ 'mat-elevation-z2': scrollTop > 200 }"
-        style="position: fixed; top: 0; z-index: 10;"
-        [style.width.px]="width"
-        *ngIf="shrinkToolbar"
-    >
-        <mat-toolbar-row>
-            <button mat-icon-button>
-                <mat-icon>menu</mat-icon>
-            </button>
-            <div fxFlex="1 1 0px" fxLayout="column" fxLayoutAlign="start" class="spacing-16">
-                <div class="mat-title" style="margin-bottom: 2px;">President</div>
-                <div class="mat-subheading-2" style="line-height: 0.6rem;">Leader of the Free World</div>
-            </div>
-            <button mat-icon-button>
-                <mat-icon>more_vert</mat-icon>
-            </button>
-        </mat-toolbar-row>
-    </mat-toolbar>
-    <mat-toolbar color="primary" class="abe" #scrollTracker>
-        <mat-toolbar-row style="justify-content: space-between;">
-            <button mat-icon-button>
-                <mat-icon>menu</mat-icon>
-            </button>
-            <button mat-icon-button>
-                <mat-icon>more_vert</mat-icon>
-            </button>
-        </mat-toolbar-row>
-        <mat-toolbar-row
-            fxLayout="column"
-            fxLayoutAlign="start"
-            style="height: auto; margin-left: 56px;"
-            [ngStyle]="{ opacity: 1 - scrollTop / 200 }"
-        >
-            <h4 class="mat-display-1">President</h4>
-            <div class="mat-headline">Commander in Chief</div>
-            <div class="mat-title">Leader of the Free World</div>
-        </mat-toolbar-row>
-        <mat-toolbar-row></mat-toolbar-row>
-    </mat-toolbar>
-
-    <mat-list>
-        <mat-list-item *ngFor="let item of list" style="padding: 16px;">
-            <mat-icon mat-list-avatar>person</mat-icon>
-            <div mat-line>{{ item.president }}</div>
-            <div mat-line>{{ item.party }}</div>
-            <div mat-line>{{ item.took_office }}</div>
-        </mat-list-item>
-    </mat-list>
-</div>
