--- src/app/pages/app-bar/app-bar.component.scss
@@ -1,22 +1,0 @@
-:host ::ng-deep .mat-list-item-content {
-    padding: 0 !important;
-}
-
-.spacing-16 {
-    margin-left: 16px;
-}
-
-.spacing-24 {
-    margin-left: 24px;
-}
-.abe {
-    //we set the background to a linear gradient so that we can have an image and a color without needing a pseudo element. If you try to fake it with transparency and no gradient the transparency will be inheirited by all the children.
-    // IE does not support background blending, so to see the image, you'll need to switch the url and the gradient
-    background: linear-gradient(rgba(0, 123, 193, 1), rgba(0, 75, 158, 1)),
-        url(https://www.biography.com/.image/ar_1:1%2Cc_fill%2Ccs_srgb%2Cg_face%2Cq_auto:good%2Cw_300/MTIwNjA4NjMzODg2NTc0MDky/abraham-lincoln-9382540-2-402.jpg) !important;
-    //use these combinations as they will be more forgiving to the original image color
-    background-repeat: no-repeat !important;
-    background-position: right !important;
-    //soft-light helps retain some truth in the original image color
-    background-blend-mode: soft-light;
-}
