--- src/app/pages/data-list/data-list.component.html
@@ -1,13 +1,0 @@
-<mat-toolbar color="primary">
-    <div>Key Value List</div>
-    <div fxFlex style="margin: auto;"></div>
-</mat-toolbar>
-<mat-list>
-    <mat-list-item *ngFor="let item of list | keyValue">
-        <div fxLayout="row" style="width: 100%;">
-            <span>{{ item.key | unCamelCase }}</span>
-            <span fxFlex></span>
-            <span>{{ item.value }}</span>
-        </div>
-    </mat-list-item>
-</mat-list>
