--- src/app/pages/data-list/data-list.component.scss
@@ -1,6 +1,0 @@
-.selected {
-    background-color: #cccccc;
-}
-mat-nav-list {
-    padding-top: 0px;
-}
