--- src/app/pages/form-validation/form-validation.component.scss
@@ -1,23 +1,0 @@
-@import '~@pxblue/colors/palette.scss';
-
-section {
-    margin: 20px;
-}
-
-mat-card {
-    padding: 20px;
-    margin-bottom: 50px;
-}
-
-mat-form-field {
-    width: 100%;
-    margin-bottom: 35px;
-}
-
-.success {
-    color: map-get($pxb-green, 500);
-}
-
-.failure {
-    color: map-get($pxb-red, 500);
-}
