--- src/app/pages/form-validation/form-validation.component.spec.ts
@@ -1,24 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { FormValidationComponent } from './form-validation.component';
-
-describe('FormValidationComponent', () => {
-    let component: FormValidationComponent;
-    let fixture: ComponentFixture<FormValidationComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            declarations: [FormValidationComponent],
-        }).compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(FormValidationComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
