+++ src/app/pages/list/data-list/data-list.component.html
@@ -1,0 +1,10 @@
+<mat-toolbar color="primary">
+    <div>Key Value List</div>
+    <div fxFlex style="margin: auto;"></div>
+</mat-toolbar>
+<mat-list>
+    <pxb-info-list-item *ngFor="let item of list | keyValue" [hidePadding]="true">
+        <span pxb-title>{{ item.key | unCamelCase }}</span>
+        <span pxb-right-content>{{ item.value }}</span>
+    </pxb-info-list-item>
+</mat-list>
