+++ src/app/pages/list/data-list/data-list.component.scss
@@ -1,0 +1,3 @@
+mat-nav-list {
+    padding-top: 0;
+}
