+++ src/app/pages/placeholder/placeholder.component.ts
@@ -1,13 +1,11 @@
 import { Component } from '@angular/core';
-import {Router} from "@angular/router";
 import {StateService} from "../../services/state.service";
 import {ViewportService} from "../../services/viewport.service";
-import * as Colors from '@pxblue/colors';
 
 @Component({
     selector: 'app-placeholder',
     template: `<div style="height: 100%; width: 100%; min-height: 100vh">
-      <mat-toolbar [style.backgroundColor]="colors.blue[500]" [style.color]="colors.white[50]">
+      <mat-toolbar color="primary" style="color: white">
         <button
           *ngIf="viewportService.isSmall()"
           mat-icon-button
@@ -22,8 +20,6 @@ import * as Colors from '@pxblue/colors';
     </div>`,
 })
 export class PlaceholderComponent {
-
-  colors = Colors;
 
   constructor(public readonly stateService: StateService, public viewportService: ViewportService) {}
